Restructured code layout
